+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Allow line and block comments inside `properties`, `perspectives` and `users` blocks
 - Parser error when using unquoted arguments containing curly braces (fixes #58)
+- Highlighting/indentation of parallel sequences (fixes #70)
 
 ## [0.3.3] - 2023-12-29